+++ mem/hp_malloc_stats.c
@@ -177,9 +177,9 @@ void hp_init_shm_statistics(struct hp_block *hpb)
 	shm_rused->flags |= STAT_NO_RESET;
 	shm_frags->flags |= STAT_NO_RESET;
 #endif
-	update_stat(shm_used, hpb->used);
-	update_stat(shm_rused, hpb->real_used);
-	update_stat(shm_frags, hpb->total_fragments);
+	update_stat(shm_used, (int)hpb->used);
+	update_stat(shm_rused, (int)hpb->real_used);
+	update_stat(shm_frags, (int)hpb->total_fragments);
 
 	LM_DBG("initializing atomic shm statistics: "
 	       "[ us: %ld | rus: %ld | frags: %ld ]\n", hpb->used, hpb->real_used, hpb->total_fragments);
